+++ src/dpr/components/_charts/chart/line/clientClass.mjs
@@ -16,7 +16,6 @@ class LineChartVisualisation extends ChartVisualisation {
 
   initSettings() {
     return {
-      options: this.setOptions(),
       toolTipOptions: this.setToolTipOptions(),
       styling: this.setDatasetStyling(),
     }
@@ -31,23 +30,11 @@ class LineChartVisualisation extends ChartVisualisation {
   }
 
   setDatasetStyling() {
-    const pallette = this.getColourPallette()
-    return pallette.map((colour) => {
-      return {
-        borderColor: colour.hex,
-        backgroundColor: colour.hex,
-        pointStyle: 'circle',
-        pointRadius: 4,
-        pointHoverRadius: 10,
-        pointHitRadius: 20,
-        datalabels: {
-          display: false,
-        },
-        segment: {
-          borderDash: (ctx) => this.setPartialStyle(ctx),
-        },
-      }
-    })
+    return {
+      segment: {
+        borderDash: (ctx) => this.setPartialStyle(ctx),
+      },
+    }
   }
 
   setToolTipOptions() {
@@ -66,24 +53,6 @@ class LineChartVisualisation extends ChartVisualisation {
           const value = data[context.dataIndex]
           ctx.setHoverValue({ label, value, legend, ctx })
           return value
-        },
-      },
-    }
-  }
-
-  setOptions() {
-    return {
-      scales: {
-        y: {
-          min: 0,
-          ticks: {
-            fontSize: 12,
-          },
-        },
-        x: {
-          ticks: {
-            fontSize: 12,
-          },
         },
       },
     }