+++ src/dpr/components/_dashboards/dashboard-visualisation/types.ts
@@ -1,7 +1,16 @@
+import z from 'zod'
 import { components } from '../../../types/api'
 import { Scorecard, ScorecardGroup } from '../scorecard/types'
 import { MatrixChartData } from '../../_charts/chart/heatmap/types'
 import { ChartDetails } from '../../../types/Charts'
+import DashboardVisualisationSchemas from './Validate'
+import { BarDefinitionMeasure } from '../../_charts/chart/bar/types'
+import { DoughnutDefinitionMeasure } from '../../_charts/chart/doughnut/types'
+import { LineDefinitionMeasure } from '../../_charts/chart/line/types'
+import { LineTimeseriesDefinitionMeasure } from '../../_charts/chart/line-timeseries/types'
+import { BarTimeseriesDefinitionMeasure } from '../../_charts/chart/bar-timeseries/types'
+import { ChartOptionsType } from '../../_charts/chart/chart-config'
+import { PartialDate } from '../../_filters/types'
 
 export interface DashboardSection {
   id: string
@@ -31,9 +40,14 @@ export interface DashboardVisualisatonCardData {
 }
 
 export interface DashboardVisualisationData {
-  type: components['schemas']['DashboardVisualisationDefinition']['type']
-  unit?: components['schemas']['DashboardVisualisationColumnDefinition']['unit']
+  type: DashboardVisualisationType
+  options?: DashboardVisualisationDataOptions
   data: DashboardVisualisationDataValues
+}
+
+interface DashboardVisualisationDataOptions {
+  height?: number
+  unit?: components['schemas']['DashboardVisualisationColumnDefinition']['unit']
   timeseries?: boolean
 }
 
@@ -41,6 +55,8 @@ export interface DashboardVisualisationDataValues {
   labels?: string[]
   datasets: DashboardVisualisationDataSet[]
   axis?: 'x' | 'y'
+  config: ChartOptionsType
+  partialDate?: PartialDate | undefined
 }
 
 export interface DashboardVisualisationDataSet {
@@ -100,3 +116,17 @@ export interface DashboardVisualisationBucket {
   max?: number | undefined
   hexColour?: string | undefined
 }
+
+export type VisualisationDefinitionType = z.infer<typeof DashboardVisualisationSchemas.DashboardVisualisationSchema>
+export type VisualisationDefinitionKey = z.infer<typeof DashboardVisualisationSchemas.DashboardVisualisationKeySchema>
+export type VisualisationDefinitionMeasure = z.infer<
+  typeof DashboardVisualisationSchemas.DashboardVisualisationMeasureSchema
+>
+
+export type ChartMeasure =
+  | BarDefinitionMeasure[]
+  | DoughnutDefinitionMeasure[]
+  | LineDefinitionMeasure[]
+  | VisualisationDefinitionMeasure[]
+
+export type TimeseriesChartMeasure = LineTimeseriesDefinitionMeasure[] | BarTimeseriesDefinitionMeasure[]