+++ test-app/mocks/mockClients/dashboards/definitions/examples/visualisations/index.js
@@ -5,12 +5,18 @@ const charts = require('./charts')
 const scorecards = require('./scorecards')
 const scorecard = require('./scorecard')
 const scorecardGroups = require('./scorecard-groups')
+const bar = require('./bar')
+const doughnut = require('./doughnut')
+const line = require('./line')
 
 module.exports = {
+  bar,
+  doughnut,
   lists,
   charts,
   scorecard,
   scorecards,
   scorecardGroups,
   matrix,
+  line,
 }